# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cosmic Webapp",
    "version": "1.0.0",
    "description": "API documentation with instructions"
  },
  "servers": [
    {
      "url": "/cosmic/"
    },
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/document": {
      "post": {
        "summary": "Upload a document to Supabase storage",
        "tags": ["Documents"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document file to upload"
                  },
                  "name": {
                    "type": "string",
                    "description": "Name of the document"
                  },
                  "task_id": {
                    "type": "integer",
                    "description": "Task ID"
                  },
                  "location_id": {
                    "type": "string",
                    "description": "Location ID"
                  },
                  "room_id": {
                    "type": "integer",
                    "description": "Room ID"
                  }
                },
                "required": ["file", "name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "description": "Path to the uploaded file"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/document/{id}": {
      "put": {
        "summary": "Update a document and replace the file",
        "tags": ["Documents"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the document to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "New document file to upload"
                  },
                  "name": {
                    "type": "string"
                  },
                  "task_id": {
                    "type": "integer"
                  },
                  "location_id": {
                    "type": "string"
                  },
                  "room_id": {
                    "type": "integer"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded"
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get a document by ID",
        "tags": ["Documents"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the document to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Document retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete document",
        "description": "Deletes a document by ID and removes the associated file from the Supabase bucket.",
        "tags": ["Documents"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the document to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Document removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Error while remove document"
          }
        }
      }
    },
    "/document/task/{task_id}": {
      "get": {
        "summary": "Get a document by task ID",
        "tags": ["Documents"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Task ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Documents retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Documents not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/documents": {
      "get": {
        "summary": "Get all documents for a specific user",
        "tags": ["Documents"],
        "security": [
          {
            "ACCESS_TOKEN": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/all_documents": {
      "get": {
        "summary": "Get all documents",
        "tags": ["Documents"],
        "parameters": [
          {
            "name": "floor_id",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "room_id",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "location_id",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/floors": {
      "post": {
        "summary": "Create floor",
        "tags": ["Floors"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "matterport_floor_id", "sequence"],
                "properties": {
                  "name": { "type": "string" },
                  "matterport_floor_id": { "type": "string" },
                  "sequence": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Floor created"
          }
        }
      },
      "get": {
        "summary": "Get all floors",
        "tags": ["Floors"],
        "responses": {
          "200": {
            "description": "List of floors"
          }
        }
      }
    },
    "/floors/{id}": {
      "get": {
        "summary": "Get floor by ID",
        "tags": ["Floors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Floor data"
          },
          "404": {
            "description": "Floor not found"
          }
        }
      },
      "put": {
        "summary": "Update floor by ID",
        "tags": ["Floors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "matterport_floor_id": { "type": "string" },
                  "sequence": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Floor updated"
          }
        }
      },
      "delete": {
        "summary": "Delete floor by ID",
        "tags": ["Floors"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Floor deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ACCESS_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "REFRESH_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Document": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Document #1"
          },
          "task_id": {
            "type": "integer",
            "example": 1
          },
          "file_url": {
            "type": "string",
            "example": "some url file"
          },
          "room_id": {
            "type": "integer",
            "example": 1
          },
          "location_id": {
            "type": "string",
            "example": 2
          },
          "user_id": {
            "type": "integer",
            "example": 1
          }
        }
      }
    }
  }
}